+++ README.md
@@ -5,3 +5,4 @@ Gets you up and running with a Frigg Application, based on Create React App.
 To read more about Frigg, [head over to our main site](https://friggframework.org)
 
 We're in early days here... please don't hesitate to write an issue, ask us questions, or even contribute!
+
