+++ packages/cfa-template/template/backend/src/workers/examples/ExamplePollWorker.js
@@ -1,6 +1,6 @@
 const moment = require('moment');
 const { Worker } = require('@friggframework/core');
-const IntegrationFactory = require('../../managers/integrations/IntegrationManager.js');
+const IntegrationFactory = require('../../managers/integrations/IntegrationManagerFactory');
 const { HaltError } = require('@friggframework/errors');
 
 const events = {
